refactor: 장르 취향조사 화면 ui를 개선한다.
