+++ android/app/src/main/java/com/digginroom/digginroom/feature/genretaste/GenreTasteBindingAdapter.kt
@@ -3,33 +3,18 @@ package com.digginroom.digginroom.feature.genretaste
 import android.graphics.Color
 import android.graphics.PorterDuff
 import android.widget.ImageView
-import androidx.annotation.DrawableRes
 import androidx.databinding.BindingAdapter
-import com.bumptech.glide.Glide
 
 object GenreTasteBindingAdapter {
 
     @JvmStatic
-    @BindingAdapter("app:src")
-    fun setImageSrc(
-        imageView: ImageView,
-        @DrawableRes
-        imageResId: Int
-    ) {
-        Glide.with(imageView.context)
-            .load(imageResId)
-            .centerCrop()
-            .into(imageView)
-    }
-
-    @JvmStatic
-    @BindingAdapter("app:hasBrightness")
-    fun setImageBrightness(
+    @BindingAdapter("app:hasDarkness")
+    fun setImageFilter(
         imageView: ImageView,
         hasBrightness: Boolean
     ) {
         if (hasBrightness) {
-            imageView.setColorFilter(Color.parseColor("#BDBDBD"), PorterDuff.Mode.MULTIPLY)
+            imageView.setColorFilter(Color.parseColor("#858585"), PorterDuff.Mode.MULTIPLY)
         } else {
             imageView.colorFilter = null
         }
